% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{facets-shiny}
\alias{facets-shiny}
\alias{facets_output}
\alias{render_facets}
\title{Shiny bindings for facets}
\usage{
facets_output(outputId, width = "100\%", height = "auto")

render_facets(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended. \code{height} will probably not
have an effect; instead, use the \code{height} parameter in
\code{\link[facets]{facets}}.}

\item{expr}{An expression that generates a facets canvas}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using facets within Shiny
applications and interactive Rmd documents.
}
