% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{facets}
\alias{facets}
\alias{facets_dive}
\alias{facets_overview}
\title{Explorer data using facets}
\usage{
facets(
  data = list(),
  type = c("facets-dive", "facets-overview"),
  width = NULL,
  height = 600,
  elementId = NULL
)

facets_dive(...)

facets_overview(...)
}
\arguments{
\item{data}{put data depends from facets type
- facets-dive - data frame (multi table is not supported yet)
- facets-overview - data frame or list of data frame as list of list}

\item{type}{one of the type of facets c("facets-dive", "facets-overview")}

\item{width}{width for html widget}

\item{height}{height for html widget by default 600}

\item{elementId}{id of the element}

\item{...}{arguments to facets wihout type}
}
\value{
the htmlwidget
}
\description{
This function generate htmlwidget object that using facets from Google Pair
}
\examples{
\donttest{
# Use predefined function or standard facets
# facets dive
facets_dive(iris)

# facets overview
facets_overview(iris)
facets_overview(list(list(data = iris, name = "iris"), list(data = cars, name = "cars")))
}
\donttest{inst/examples/app.R}
}
\seealso{
\href{https://pair-code.github.io/facets/}{Facets Google pair website}
}
