% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facetsr.R
\name{inst_path}
\alias{inst_path}
\title{Access files in the current app}
\usage{
inst_path(...)
}
\arguments{
\item{...}{Character vector specifying directory and or file to
point to inside the current package.}
}
\description{
Access files in the current app
}
\examples{
\dontrun{
inst_path("")
}
}
